feat: 歩数抽出のクエリパラメータを組み立てる処理を実装
